fix: send error to execute:failure webhook
