+++ src/utils/paypal/agreements/error/status.js
@@ -1,11 +1,11 @@
 const CODE_AGREEMENT_STATUS_FORBIDDEN = 'agreement-status-forbidden';
 
 class AgreementStatusError extends Error {
-  constructor(agreementId, owner, status, creatorTaskId) {
+  constructor(agreementId, owner, status, creatorTaskId, token) {
     super(`Agreement "${agreementId}" has status "${status}"`);
 
     this.code = CODE_AGREEMENT_STATUS_FORBIDDEN;
-    this.params = { agreementId, owner, status, creatorTaskId };
+    this.params = { agreementId, owner, status, creatorTaskId, token };
   }
 
   getHookErrorData() {
